10/07/2025 6:28PM
I added the CPU.java, and the MainMemory.java.
My thoughts: I plan to look through all the code and strategise what to do moving forward.
Plan: I plan to add the 3 files needed, and see what to code in them.
This message might be repeated on github. The IDE and github were not connecting correctly, this is a test.