10/14/2025 1:35PM
I have started working on the encryption.java file. So far I have worked on the Encryption file, and I am planning on implementing a cipher.

My thoughts: I am implementing a cipher to encrypt the text. I now need to work on decrypting the text; the decrypt portion of the code.

Plan: Right now, the plan has shifted to work on each file one at a time. I plan to get the basics down, and refer back to the pseudocode I have written down. The pseudocode may end up being more of what I need the code to accomplish, rather than follow it down to a tee.